initial test setup
This calculation certainly has a lot of moving pieces to it
It's also (so far) foreign to me because I'm young enough that
I've not had to give much thought to Social Security, indeed
I don't know if it'll exist in it's current format when I'm older.

Anyway, this commit and the next few will be a few 'spike' commits
where I'm exploring a few different ideas and avenues to
try to wrap my head around an approach to the solution.